query-service/pyproject.toml
[tool.poetry]
name = "query-service"
version = "1.4.1"
description = "Query service for SaaS B2B using Clean Architecture, PyMilvus, Haystack & Advanced RAG with remote embeddings, reranking, and sparse search."
authors = ["Nyro <dev@atenex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
gunicorn = "^21.2.0"
pydantic = {extras = ["email"], version = "^2.6.4"} 
pydantic-settings = "^2.2.1"
httpx = ">=0.27.0,<1.0.0" 
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tenacity = "^8.2.3"
structlog = "^24.1.0"
haystack-ai = "^2.0.1" 
pymilvus = "==2.5.3" 
numpy = "1.26.4" 
tiktoken = "^0.9.0"
# google-generativeai REMOVED

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"